modules/sound: support wireplumber 0.5.2 and above
With the release of the 2.x series, asahi-audio is now compatible with
wireplumber 0.5.2 and above:

> The stable software DSP API appears in WirePlumber 0.5.2.
https://github.com/AsahiLinux/asahi-audio/commit/29ec1056c18193ffa09a990b1b61ed273e97fee6

We remove the wireplumber downgrade and assert this new minimum wireplumber version
present in nixpkgs. In addition, we add the asahi-audio share/ path as a data
directory to the wireplumber systemd user unit, as wireplumber now only looks
for Lua scripts in datadirs only.